fix: parse inputs and left them untouched in UI builder
